requirements.txt
addict==2.4.0
aggdraw==1.3.19
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.0
cycler==0.12.1
docopt==0.6.2
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
groundingdino-py==0.4.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
psd-tools==1.10.9
pycocotools==2.0.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
supervision==0.6.0
sympy==1.14.0
tifffile==2024.8.30
timm==1.0.19
tokenizers==0.21.4
tomli==2.2.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm==4.67.1
transformers==4.55.4
typing_extensions==4.15.0
urllib3==2.5.0
yapf==0.43.0
zipp==3.23.0
